# cmd/api/docs/swagger.yaml
definitions:
...
  models.AdminLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Show Times is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Show Times
  version: 1.0.0
paths:
  /admin/block:
    put:
      consumes:
      - application/json
      description: Blocks a user based on the provided ID.
      parameters:
      - description: User ID to block
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User blocked successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to block user
          schema:
            $ref: '#/definitions/response.Response'
      summary: Block a user
      tags:
      - User Management
...
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: Retrieves dashboard information for admin.
      produces:
      - application/json
      responses:
        "200":
          description: Admin dashboard retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or unable to retrieve dashboard
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieve admin dashboard information
      tags:
      - Admin Dashboard
  /admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate an admin and get an access token.
      parameters:
      - description: Admin credentials for login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Admin login successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or incorrect format
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: cannot authenticate user'
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin login
      tags:
      - Admin Authentication
...
  /admin/salesreport:
    get:
      consumes:
      - application/json
      description: Retrieves sales report for the specified time period.
      parameters:
      - description: Time period for sales report (e.g., weekly, monthly)
        in: query
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales report retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Unable to retrieve sales report
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieve sales report for a specific time period
      tags:
      - Sales Reports
  /admin/salesreport/daterange:
    get:
      consumes:
      - application/json
      description: Retrieves sales report based on the provided start and end dates.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales report retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Unable to retrieve sales report
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieve sales report by date range
      tags:
      - Sales Reports
  /admin/unblock:
    patch:
      consumes:
      - application/json
      description: Unblocks a user based on the provided ID.
      parameters:
      - description: User ID to unblock
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User unblocked successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or unable to unblock user
          schema:
            $ref: '#/definitions/response.Response'
      summary: Unblock a user
      tags:
      - User Management
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retrieves users based on the provided page number.
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request or unable to retrieve users
          schema:
            $ref: '#/definitions/response.Response'
      summary: Retrieve users with pagination
      tags:
      - User Management
...
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"